test: name every test file after the module it tests, and gate it
The convention from the PR 269 sweep had no gate, and six orphans crept
back in. The wiring tests dissolve into cmd/server/main_test.go, the
See-also guard into tool_manifest_test.go, the catalog memoization
helper and the text-name guard into action_catalog_test.go, and the two
qualified names move onto their module: packages_stream_unix_test.go
and kind_integration_test.go (external test package forced by the
resources import cycle, so it cannot share kind_test.go's name).

cmd/audit_test_names -check-files is the codified gate, wired as
make check-test-file-names in CI: a basename must match a module file,
with export_test.go, build-constrained qualifiers, external-package
qualifiers, and test/e2e as the recorded exemptions. Affected packages
verified under -shuffle=on.
