+++ README.md
@@ -446,18 +446,18 @@ and is never logged. Full details: [PRIVACY.md](PRIVACY.md).
 
 | Category                 |     Files |       Lines |
 | ------------------------ | --------: | ----------: |
-| Source (`.go`, non-test) |     1,007 |     208,612 |
-| Unit tests (`_test.go`)  |       565 |     314,637 |
+| Source (`.go`, non-test) |     1,008 |     208,790 |
+| Unit tests (`_test.go`)  |       561 |     314,599 |
 | End-to-end tests         |       212 |      56,627 |
-| **Total**                | **1,784** | **579,876** |
+| **Total**                | **1,781** | **580,016** |
 
 ### Functions
 
 | Category                        |  Count |
 | ------------------------------- | -----: |
-| Source functions                |  7,842 |
+| Source functions                |  7,850 |
 | — exported (public)             |  2,707 |
-| — unexported (private)          |  5,135 |
+| — unexported (private)          |  5,143 |
 | Unit test functions (`TestXxx`) | 11,788 |
 | Subtests (`t.Run(...)`)         |  3,122 |
 | End-to-end test functions       |    539 |
@@ -468,17 +468,17 @@ and is never logged. Full details: [PRIVACY.md](PRIVACY.md).
 | ---------------------------------- | -------------------------: |
 | Test lines vs source lines         | 1.51× more tests than code |
 | Average source file length         |                 ~207 lines |
-| Average test file length           |                 ~556 lines |
-| Comment lines in source            |  28,256 (~13.5% of source) |
+| Average test file length           |                 ~560 lines |
+| Comment lines in source            |  28,296 (~13.6% of source) |
 | Test functions per source function |                       1.5× |
 
 ### Code patterns
 
 | Pattern                            | Count |
 | ---------------------------------- | ----: |
-| `if err != nil` checks             | 6,655 |
+| `if err != nil` checks             | 6,658 |
 | `defer` statements                 | 1,020 |
-| `struct` types defined             | 2,759 |
+| `struct` types defined             | 2,760 |
 | `//nolint` suppressions            |   250 |
 | `TODO` / `FIXME` / `HACK` comments |     2 |
 
@@ -501,7 +501,7 @@ and is never logged. Full details: [PRIVACY.md](PRIVACY.md).
 
 | Fact                                 | Value                                                                                                |
 | ------------------------------------ | ---------------------------------------------------------------------------------------------------- |
-| Source code printed at 55 lines/page | ~3,792 pages of A4                                                                                   |
+| Source code printed at 55 lines/page | ~3,796 pages of A4                                                                                   |
 | Source lines mentioning `"gitlab"`   | 12,779 (impossible to avoid)                                                                         |
 | Longest function name in source      | `assertDynamicCompatibilityPolicyOwnedByActionCompat` (51 chars)                                     |
 | Longest test function name           | `TestRequiredMissingAndUnknownParamNames_SchemaValidation_ReturnsSortedMissingAndUnknown` (87 chars) |
